% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BeeRs-package.R
\docType{package}
\name{BeeRs-package}
\alias{BeeRs}
\alias{BeeRs-package}
\title{BeeRs: BeeRs production in USA - interactive visualisation}
\description{
Import the data and Launch a shiny app.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pmoizerv/BeeRs}
  \item Report bugs at \url{https://github.com/pmoizerv/BeeRs/issues}
}

}
\author{
\strong{Maintainer}: Panagiotis Zervakis \email{panagiotis.zervakis@su.se}

Other contributors:
\itemize{
  \item Kinlan Jan \email{kinlan.jan@su.se} [contributor]
}

}
\keyword{internal}
