% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beers.R
\docType{package}
\name{beers}
\alias{beers}
\alias{beers-package}
\title{beers: A package for computing Beers ordinary and modified interpolation}
\description{
beers: A package for computing Beers ordinary and modified interpolation
}
