% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beers_dat.R
\docType{data}
\name{beers_dat}
\alias{beers_dat}
\title{Beers production data}
\format{
\subsection{\code{beers_dat}}{

A data frame with 7,240 rows and 60 columns:
\describe{
\item{state}{state name}
\item{year}{Year}
\item{state_full}{Full name of state}
\item{Permises}{Volume in L for consumption on permises}
\item{BottlesCans}{Volume in L for comsumption in bottles and cans}
\item{KegsBarrels}{Volume in L for comsumption in Kegs and barrels}
...
}
}
}
\source{
\url{https://www.kaggle.com/datasets/jessemostipak/beer-production}
}
\usage{
beers_dat
}
\description{
An overview of beers production by state in the USA
Report ...
}
\keyword{datasets}
