% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beers_functions.R
\name{beers_int_modified}
\alias{beers_int_modified}
\title{Beers modified algorithm for interpolation.}
\usage{
beers_int_modified(points)
}
\arguments{
\item{points}{A list of at least 6 numbers.}
}
\value{
A list where each original number exists with four interpolated points between each.
}
\description{
Beers modified algorithm for interpolation.
}
\examples{
beers_int_modified(1:6)
}
