% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beers-r6.R
\name{beers_ordinary}
\alias{beers_ordinary}
\title{Beers ordinary algorithm for interpolation.}
\usage{
beers_ordinary(points)
}
\arguments{
\item{points}{A list of at least 6 numbers.}
}
\value{
A list where each original number exists with four interpolated points between each.
}
\description{
Beers ordinary algorithm for interpolation.
}
\examples{
beers_ordinary(c(1:6)) returns 1.0 1.2 1.4 1.6 1.8 2.0 2.2 ... 5.8 6.0

}
