% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beers_functions.R
\name{beers_sub_ordinary}
\alias{beers_sub_ordinary}
\alias{beers_sub_modified}
\title{Beers subdivision, using ordinary or modified method.}
\usage{
beers_sub_ordinary(points)

beers_sub_modified(points)
}
\arguments{
\item{points}{A list of at least 5 numbers, eg, populations by 5-year age-band.}
}
\value{
For subdivision: a list 5 times as long as the original, with each point subdivided into 5.
}
\description{
With the ordinary method, each set of 5 subdivided values will always sum to the original data point given.
With the modified method, some smoothing occurs, and this property is only true for the first and last data point.
}
\examples{

# Subdivide population of UK (2015), ages 0-4, 5-9, 10-14, 15-19, 20-24 into single years.

beers_sub_ordinary(c(4042918, 3927745, 3529200, 3779712, 4174572))
beers_sub_modified(c(4042918, 3927745, 3529200, 3779712, 4174572))
}
