% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beers_functions.R
\name{beers_sub_ordinary}
\alias{beers_sub_ordinary}
\title{Beers ordinary algorithm for subdivision.}
\usage{
beers_sub_ordinary(points)
}
\arguments{
\item{points}{A list of at least 5 numbers.}
}
\value{
A list 5 times as long as the original, with each original entry subdivided into 5.
}
\description{
Beers ordinary algorithm for subdivision.
}
\examples{
beers_sub_ordinary(1:6)
}
