% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_app.R
\name{shiny_lauch}
\alias{shiny_lauch}
\title{Launch a shiny map}
\usage{
shiny_lauch(N = 0, ...)
}
\arguments{
\item{N}{Number of beers}

\item{...}{Elipses}
}
\value{
shiny map
}
\description{
Launch a shiny map
}
