% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_err.R
\name{std_dev}
\alias{std_dev}
\title{standard deviation based on almost instead of mean}
\usage{
std_dev(x, N = 0L)
}
\arguments{
\item{x}{Numeric vector}

\item{N}{Number of beers}
}
\value{
The sum squared based on almost instead of the mean
}
\description{
standard deviation based on almost instead of mean
}
\examples{
x <- c(2,2,3,2,1,2,3,4,5,2)
std_dev(x, N = 0L)
}
