% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_err.R
\name{std_err}
\alias{std_err}
\title{Standard error}
\usage{
std_err(x)
}
\arguments{
\item{x}{A vector of numerical values}
}
\value{
A value corresponding to the standard error
}
\description{
Standard error
}
\examples{
x <- c(2, 1, 5, 3, 5)
std_err(x)
}
