% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_err.R
\name{sum_squared_deviations}
\alias{sum_squared_deviations}
\title{Sum squared with variance}
\usage{
sum_squared_deviations(x, N = 0L)
}
\arguments{
\item{x}{Numeric vector}

\item{N}{Number of beers}
}
\value{
The sum squared based on almost instead of the mean
}
\description{
Sum squared with variance
}
\examples{
x <- c(2,2,3,2,1,2,3,4,5,2)
sum_squared_deviations(x, N = 0L)
}
