% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beers_functions.R
\name{beers_int_ordinary}
\alias{beers_int_ordinary}
\alias{beers_int_modified}
\title{Beers interpolation, using ordinary or modified method.}
\usage{
beers_int_ordinary(points)

beers_int_modified(points)
}
\arguments{
\item{points}{A list of at least 6 numbers, eg, populations over time.}
}
\value{
A list with four extra interpolated points between each pair of given points.
}
\description{
Add 4 interpolated points between each pair of given values.
With the ordinary method, the data points given will appear, unchanged, in the interpolated list.
With the modified method, some smoothing occurs, and only the first and last points given are guaranteed to appear unchanged in the interpolated list.
}
\examples{
# Interpolate population of UK (1950, 1955, 1960, 1965, 1970, 1975) to yearly points

beers_int_ordinary(c(50616014, 51123707, 52433157, 54303107, 55634935, 56211947))
beers_int_modified(c(50616014, 51123707, 52433157, 54303107, 55634935, 56211947))
}
